+++ cn5X_jog.py
@@ -141,7 +141,7 @@ class dlgJog(QObject):
     myWidth = self.__dlg.geometry().width()
     myHeight = self.__dlg.geometry().height()
     self.__dlg.setFixedSize(self.__dlg.geometry().width(),self.__dlg.geometry().height())
-    self.__dlg.move(ParentX + ((ParentWidth - myWidth) / 2),ParentY + ((ParentHeight - myHeight) / 2),)
+    self.__dlg.move(ParentX + int((ParentWidth - myWidth) / 2),ParentY + int((ParentHeight - myHeight) / 2),)
     self.__dlg.setWindowFlags(Qt.Dialog | Qt.Tool | Qt.WindowStaysOnTopHint)
     
     # Mise à jour de la vitesse de déplacement:
